🦆 Migrate library sort from hooks to redux
