+++ client/src/features/library/librarySlice.ts
@@ -1,4 +1,4 @@
-import { createSlice } from '@reduxjs/toolkit';
+import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 
 interface Item {
   img: string;
@@ -10,17 +10,47 @@ interface Item {
 
 interface LibraryState {
   items: Item[];
+  sortBy: 'alphabetical' | 'recentlyAdded';
 }
 
 const initialState: LibraryState = {
   items: [],
+  sortBy: 'alphabetical',
 };
+
+const sortItems = (
+  items: Item[],
+  sortBy: 'alphabetical' | 'recentlyAdded',
+): Item[] =>
+  [...items].sort((a, b) => {
+    const pinComparison = +b.isPinned - +a.isPinned;
+    if (pinComparison !== 0) return pinComparison;
+
+    if (sortBy === 'alphabetical') {
+      return a.name.localeCompare(b.name);
+    } else if (sortBy === 'recentlyAdded') {
+      return +new Date(b.createdAt) - +new Date(a.createdAt);
+    }
+
+    return 0;
+  });
 
 const librarySlice = createSlice({
   name: 'library',
   initialState,
-  reducers: {},
+  reducers: {
+    setLibraryItems: (state, action: PayloadAction<[]>) => {
+      state.items = sortItems(action.payload, 'alphabetical');
+    },
+    sortLibraryItems: (
+      state,
+      action: PayloadAction<'alphabetical' | 'recentlyAdded'>,
+    ) => {
+      state.sortBy = action.payload;
+      state.items = sortItems(state.items, state.sortBy);
+    },
+  },
 });
 
-export const {} = librarySlice.actions;
+export const { setLibraryItems, sortLibraryItems } = librarySlice.actions;
 export default librarySlice.reducer;
