+++ client/src/layout/Library/Library.jsx
@@ -3,7 +3,6 @@ import LibraryHeader from './LibraryHeader/LibraryHeader.jsx';
 import LibraryList from './LibraryList/LibraryList.jsx';
 import { useState } from 'react';
 import LibrarySearch from './LibrarySearch/LibrarySearch.tsx';
-import { LibrarySortProvider } from '../../hooks/useLibrarySort.tsx';
 
 const Library = () => {
   const [isCollapsed, setIsCollapsed] = useState(false);
@@ -13,15 +12,13 @@ const Library = () => {
   };
 
   return (
-    <LibrarySortProvider>
-      <div
-        className={`${styles.library} ${isCollapsed ? styles.libraryCollapsed : ''}`}
-      >
-        <LibraryHeader isCollapsed={isCollapsed} onCollapse={handleCollapse} />
-        {!isCollapsed && <LibrarySearch />}
-        <LibraryList />
-      </div>
-    </LibrarySortProvider>
+    <div
+      className={`${styles.library} ${isCollapsed ? styles.libraryCollapsed : ''}`}
+    >
+      <LibraryHeader isCollapsed={isCollapsed} onCollapse={handleCollapse} />
+      {!isCollapsed && <LibrarySearch />}
+      <LibraryList />
+    </div>
   );
 };
